Drop old major versions
This should cause a drop in code coverage, producing a lot of dead code.
Subsequent changes removing said dead code should not impact the
coverage in terms of number of lines covered, and would be way too
substantial to be reviewed in one go.
Support for Symfony LTS is kept.
I propose that we work on removing things like support for YML mapping
separately.