Remove useless test
On the 3.0.x branch I removed the part about auto_generate_proxy_classes
(because it's deprecated), and inadvertenly removed some of the sense
this test had (testing precedence of imports, regardless of whether the
setting is deprecated or not).

In 01c714eef7d5c5515d664b1763af186ee7abd97b, another part of the meaning
was lost (also because of a deprecation).

Now I should ideally find non-deprecated settings and test precedence
again. However, I do not see how we are testing the extension itself
here. To me, it looks like we are testing a Symfony feature. Let us see
if there is any drop in code coverage.