Inline parent class code
In order to be make it easier to make changes in the metadata discovery
mechanism, let us avoid having to change code in 2 different
repositories and inline the code of the AbstractDoctrineExtension of the
Symfony bridge here. This class will be deprecated and ultimately,
removed.

One of the methods was protected and not used, and I did not copy it
loadObjectManagerCacheDriver().

No type hints have been added for the sake of backward compatibility,
but there is still a breaking change: the extension no longer pass an
instanceof AbstractDoctrineExtension assertion, and I think that's fine.